Fix floating-point/double conversion error

+++ Multiplex/Utilities/MPXGeometry.h
@@ -11,7 +11,7 @@
 NS_INLINE CGFloat MPXRoundedPixelValueForView(NSView *view, CGFloat value)
 {
     CGFloat scale = view.window.screen.backingScaleFactor;
-    return roundf(value * scale) / scale;
+    return round(value * scale) / scale;
 }
 
 NS_INLINE CGRect MPXRoundedValueRectForView(CGRect rect, NSView *view)
